+++ src/tests/end-to-end/tests/popup/launchpad.test.ts
@@ -1,13 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { Browser } from '../../common/browser';
+import { Browser, TargetPageInfo } from '../../common/browser';
 import { launchBrowser } from '../../common/browser-factory';
 import { CommonSelectors } from '../../common/element-identifiers/common-selectors';
 import { popupPageElementIdentifiers } from '../../common/element-identifiers/popup-page-element-identifiers';
 import { enableHighContrast } from '../../common/enable-high-contrast';
 import { Page } from '../../common/page';
 import { scanForAccessibilityIssues } from '../../common/scan-for-accessibility-issues';
-import { setupNewTargetPage, TargetPageInfo } from '../../common/setup-new-target-page';
 
 describe('Launch Pad', () => {
     describe('Normal mode', () => {
@@ -17,7 +16,7 @@ describe('Launch Pad', () => {
 
         beforeAll(async () => {
             browser = await launchBrowser({ suppressFirstTimeDialog: true });
-            targetPageInfo = await setupNewTargetPage(browser);
+            targetPageInfo = await browser.setupNewTargetPage();
             popupPage = await browser.newExtensionPopupPage(targetPageInfo.tabId);
             await popupPage.bringToFront();
             await popupPage.waitForSelector(popupPageElementIdentifiers.launchPad);
@@ -46,7 +45,7 @@ describe('Launch Pad', () => {
 
         beforeAll(async () => {
             browser = await launchBrowser({ suppressFirstTimeDialog: true });
-            targetPageInfo = await setupNewTargetPage(browser);
+            targetPageInfo = await browser.setupNewTargetPage();
             const detailsViewPage = await browser.newExtensionDetailsViewPage(targetPageInfo.tabId);
             await enableHighContrast(detailsViewPage);
